Add support for praga (#9)
This PR adds support for running [praga](https://github.com/cocreators-ee/praga) for authentication as a separate service in addition to frontend and backend.

Some minor other things, like updates of some pre-commit hooks and checking if the nginx config files exist before processing them through pase-template.

+++ src/starter.py
@@ -103,13 +103,14 @@ async def setup_nginx() -> None:
 
     for f in ["/etc/nginx/conf.d/default.conf", "/etc/nginx/dataspace-headers.conf"]:
         path = Path(f)
-        final_config = await get_output(["parse-template", f])
-        path.write_bytes(final_config)
-        logger.info(f"{f} is generated")
+        if path.exists():
+            final_config = await get_output(["parse-template", f])
+            path.write_bytes(final_config)
+            logger.info(f"{f} is generated")
 
-        if os.environ.get("NGINX_DEBUG"):
-            logger.info(f"=== {f} ===")
-            logger.info(path.read_text("utf8") + "\n")
+            if os.environ.get("NGINX_DEBUG"):
+                logger.info(f"=== {f} ===")
+                logger.info(path.read_text("utf8") + "\n")
 
 
 async def start_service(svc: Service, extra_env: Optional[dict] = None) -> Process:
@@ -175,6 +176,7 @@ class Runner:
         backend: Optional[Service],
         frontend: Optional[Service],
         nginx: Optional[Service],
+        praga: Optional[Service],
     ):
         # Handle Ctrl+C gracefully by stopping all running services
         loop = asyncio.get_running_loop()
@@ -196,6 +198,10 @@ class Runner:
             prerequisites.append(
                 wait_for_service(frontend, frontend.timeout, "frontend")
             )
+
+        if praga:
+            procs.append(await start_service(praga))
+            prerequisites.append(wait_for_service(praga, praga.timeout, "praga"))
 
         if nginx:
             prerequisites.append(setup_nginx())
